mistralrs_server_core

Module openapi_doc

OpenAPI doc functionality.

Functions

get_openapi_doc

This is used to generate the OpenAPI docs. The mistral.rs server router will include these by default, but if you’re including the mistral.rs server core into another project, you can generate the OpenAPI docs separately to merge with the other project OpenAPI docs.