mistralrs_server_core::responses_types

Module items

Input and output item types for the OpenResponses API.

Structs

MessageItemParam

Message item parameter for input

Enums

InputItem

Input item types for the OpenResponses API.

MessageContentParam

Message content that can be a string or array of content parts

OutputItem

Output item types for the OpenResponses API

ResponsesInput

Input that can be a string, array of input items, or array of messages

TaggedInputItem

Tagged input item types (with explicit “type” field)