mistralrs_server_core

Module types

Custom types used in mistral.rs server core.

Type Aliases

ExtractedMistralRsState

This is the SharedMistralRsState that has been extracted for an axum handler.

OnChunkCallback

A callback function that processes streaming response chunks before they are sent to the client.

OnDoneCallback

A callback function that is executed when the streaming response completes.

SharedMistralRsState

This is the underlying instance of mistral.rs.